renamed project name walk through (Job Odyssey)
