+++ client/android/app/src/main/java/com/jobodyssey/MainActivity.java
@@ -1,4 +1,4 @@
-package com.client;
+package com.jobodyssey;
 
 import com.facebook.react.ReactActivity;
 
@@ -10,6 +10,6 @@ public class MainActivity extends ReactActivity {
    */
   @Override
   protected String getMainComponentName() {
-    return "client";
+    return "JobOdyssey";
   }
 }
