+++ client/android/app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">client</string>
+    <string name="app_name">Job Odyssey</string>
 </resources>
