+++ client/android/settings.gradle
@@ -1,3 +1,3 @@
-rootProject.name = 'client'
+rootProject.name = 'JobOdyssey'
 apply from: file("../node_modules/@react-native-community/cli-platform-android/native_modules.gradle"); applyNativeModulesSettingsGradle(settings)
 include ':app'
