+++ client/app.json
@@ -1,4 +1,4 @@
 {
-  "name": "client",
-  "displayName": "client"
+  "name": "JobOdyssey",
+  "displayName": "Job Odyssey"
 }
