+++ client/ios/JobOdyssey.xcodeproj/xcshareddata/xcschemes/JobOdyssey-tvOS.xcscheme
@@ -14,10 +14,10 @@
             buildForAnalyzing = "YES">
             <BuildableReference
                BuildableIdentifier = "primary"
-               BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-               BuildableName = "client.app"
-               BlueprintName = "client"
-               ReferencedContainer = "container:client.xcodeproj">
+               BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
+               BuildableName = "JobOdyssey-tvOS.app"
+               BlueprintName = "JobOdyssey-tvOS"
+               ReferencedContainer = "container:JobOdyssey.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -32,10 +32,10 @@
             skipped = "NO">
             <BuildableReference
                BuildableIdentifier = "primary"
-               BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "clientTests.xctest"
-               BlueprintName = "clientTests"
-               ReferencedContainer = "container:client.xcodeproj">
+               BlueprintIdentifier = "2D02E48F1E0B4A5D006451C7"
+               BuildableName = "JobOdyssey-tvOSTests.xctest"
+               BlueprintName = "JobOdyssey-tvOSTests"
+               ReferencedContainer = "container:JobOdyssey.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -54,10 +54,10 @@
          runnableDebuggingMode = "0">
          <BuildableReference
             BuildableIdentifier = "primary"
-            BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "client.app"
-            BlueprintName = "client"
-            ReferencedContainer = "container:client.xcodeproj">
+            BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
+            BuildableName = "JobOdyssey-tvOS.app"
+            BlueprintName = "JobOdyssey-tvOS"
+            ReferencedContainer = "container:JobOdyssey.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </LaunchAction>
@@ -71,10 +71,10 @@
          runnableDebuggingMode = "0">
          <BuildableReference
             BuildableIdentifier = "primary"
-            BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "client.app"
-            BlueprintName = "client"
-            ReferencedContainer = "container:client.xcodeproj">
+            BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
+            BuildableName = "JobOdyssey-tvOS.app"
+            BlueprintName = "JobOdyssey-tvOS"
+            ReferencedContainer = "container:JobOdyssey.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
