+++ client/ios/JobOdyssey.xcodeproj/xcshareddata/xcschemes/JobOdyssey.xcscheme
@@ -14,10 +14,10 @@
             buildForAnalyzing = "YES">
             <BuildableReference
                BuildableIdentifier = "primary"
-               BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-               BuildableName = "client-tvOS.app"
-               BlueprintName = "client-tvOS"
-               ReferencedContainer = "container:client.xcodeproj">
+               BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
+               BuildableName = "JobOdyssey.app"
+               BlueprintName = "JobOdyssey"
+               ReferencedContainer = "container:JobOdyssey.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -32,10 +32,10 @@
             skipped = "NO">
             <BuildableReference
                BuildableIdentifier = "primary"
-               BlueprintIdentifier = "2D02E48F1E0B4A5D006451C7"
-               BuildableName = "client-tvOSTests.xctest"
-               BlueprintName = "client-tvOSTests"
-               ReferencedContainer = "container:client.xcodeproj">
+               BlueprintIdentifier = "00E356ED1AD99517003FC87E"
+               BuildableName = "JobOdysseyTests.xctest"
+               BlueprintName = "JobOdysseyTests"
+               ReferencedContainer = "container:JobOdyssey.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -54,10 +54,10 @@
          runnableDebuggingMode = "0">
          <BuildableReference
             BuildableIdentifier = "primary"
-            BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-            BuildableName = "client-tvOS.app"
-            BlueprintName = "client-tvOS"
-            ReferencedContainer = "container:client.xcodeproj">
+            BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
+            BuildableName = "JobOdyssey.app"
+            BlueprintName = "JobOdyssey"
+            ReferencedContainer = "container:JobOdyssey.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </LaunchAction>
@@ -71,10 +71,10 @@
          runnableDebuggingMode = "0">
          <BuildableReference
             BuildableIdentifier = "primary"
-            BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-            BuildableName = "client-tvOS.app"
-            BlueprintName = "client-tvOS"
-            ReferencedContainer = "container:client.xcodeproj">
+            BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
+            BuildableName = "JobOdyssey.app"
+            BlueprintName = "JobOdyssey"
+            ReferencedContainer = "container:JobOdyssey.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
