+++ client/ios/JobOdyssey/Images.xcassets/AppIcon.appiconset/Contents.json
