+++ client/ios/JobOdysseyTests/JobOdysseyTests.m
@@ -7,11 +7,11 @@
 #define TIMEOUT_SECONDS 600
 #define TEXT_TO_LOOK_FOR @"Welcome to React"
 
-@interface clientTests : XCTestCase
+@interface JobOdysseyTests : XCTestCase
 
 @end
 
-@implementation clientTests
+@implementation JobOdysseyTests
 
 - (BOOL)findSubviewInView:(UIView *)view matching:(BOOL(^)(UIView *view))test
 {
