+++ client/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "client",
+  "name": "JobOdyssey",
   "version": "0.0.1",
   "private": true,
   "scripts": {
